scraper/requirements.txt
SQLAlchemy
psycopg-binary
pandas
numpy
requests
beautifulsoup4
RapidFuzz
python-dotenv
tqdm
PyYAML
